Add the index file in navigation folder and project runs fine
