+++ src/Redux/actions/auth-action/auth-action.js
@@ -1,4 +1,4 @@
-import * as TYPES from '../../../types/auth-types/auth-types';
+import * as TYPES from '../../types/auth-types/auth-types';
 
 const setUserData = payload => {
   return {type: TYPES.LOGIN_REQUEST_REQUEST, payload};
